+++ src/common/avc_hevc/util.h
@@ -1,0 +1,21 @@
+/** AVC & HEVC video helper functions
+
+   mkvmerge -- utility for splicing together matroska files
+   from component media subtypes
+
+   Distributed under the GPL v2
+   see the file COPYING for details
+   or visit https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+
+   \author Written by Moritz Bunkus <moritz@bunkus.org>.
+*/
+
+#pragma once
+
+#include "common/common_pch.h"
+
+namespace mtx::avc_hevc {
+
+bool might_be_avc_or_hevc(memory_c const &buffer);
+
+}